Everything works perfectly, I can change a little things to make it look better and more creative

+++ lib/components/constants.dart
@@ -13,7 +13,7 @@ int kLowerContainer = 0xFFEB1555;
 double kSliderValue = 1;
 
 
-int weight = 45;
+double weight = 45;
 
 int age = 18;
